Line Chart all data showing

+++ app/src/main/java/com/example/exercisetracker/data/LineChartData.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class LineChartData(
     val startTime: List<Long>,
-    val totalKM: List<Double>,
+    val totalKM: List<Float>,
     val totalTime: List<Long>,
-    val averageSpeed: List<Double>
+    val averageSpeed: List<Float>
 ) : Parcelable
